backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
openai==1.30.1
google-generativeai>=0.7.2
httpx==0.27.0
apscheduler==3.10.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
